+++ packages/blade/src/components/Typography/Heading/Heading.stories.tsx
@@ -88,19 +88,13 @@ const WithMixedColorsTemplate: ComponentStory<typeof HeadingComponent> = () => {
     <Box>
       <HeadingComponent>
         Supercharge your business with the all‑powerful{' '}
-        <HeadingComponent
-          as="span"
-          color="feedback.information.action.text.primary.default.lowContrast"
-        >
+        <HeadingComponent as="span" color="brand.primary.600">
           Payment Gateway
         </HeadingComponent>
       </HeadingComponent>
       <HeadingComponent marginTop="spacing.5">
         Start accepting{' '}
-        <HeadingComponent
-          as="span"
-          color="feedback.information.action.text.primary.default.lowContrast"
-        >
+        <HeadingComponent as="span" color="feedback.text.information.lowContrast">
           payments
         </HeadingComponent>{' '}
         at just 2% <Sup>*</Sup>